Fix flaky test cases

+++ integration-test-standard/src/test/java/com/ericsson/bss/cassandra/ecaudit/integration/standard/ITVerifyAudit.java
@@ -97,10 +97,10 @@ public class ITVerifyAudit
         "ALTER ROLE cassandra WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/system_schema' }"));
 
         session.execute(new SimpleStatement(
-        "CREATE KEYSPACE ecks WITH REPLICATION = {'class' : 'SimpleStrategy', 'replication_factor' : 1} AND DURABLE_WRITES = false"));
-        session.execute(new SimpleStatement("CREATE TABLE ecks.ectbl (partk int PRIMARY KEY, clustk text, value text)"));
+        "CREATE KEYSPACE ecks_itva WITH REPLICATION = {'class' : 'SimpleStrategy', 'replication_factor' : 1} AND DURABLE_WRITES = false"));
+        session.execute(new SimpleStatement("CREATE TABLE ecks_itva.ectbl (partk int PRIMARY KEY, clustk text, value text)"));
         session.execute(new SimpleStatement(
-        "CREATE TABLE ecks.ectypetbl (partk int PRIMARY KEY, v0 text, v1 ascii, v2 bigint, v3 blob, v4 boolean, "
+        "CREATE TABLE ecks_itva.ectypetbl (partk int PRIMARY KEY, v0 text, v1 ascii, v2 bigint, v3 blob, v4 boolean, "
         + "v5 date, v6 decimal, v7 double, v8 float, v9 inet, v10 int, v11 smallint, v12 time, v13 timestamp, "
         + "v14 uuid, v15 varchar, v16 varint)"));
 
@@ -114,17 +114,17 @@ public class ITVerifyAudit
         session.execute(new SimpleStatement(
         "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/system_schema'}"));
         session.execute(new SimpleStatement(
-        "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/ecks/ectbl'}"));
+        "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/ecks_itva/ectbl'}"));
         session.execute(new SimpleStatement(
         "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/nonexistingks'}"));
         session.execute(new SimpleStatement(
-        "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/ecks/nonexistingtbl'}"));
+        "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data/ecks_itva/nonexistingtbl'}"));
         session.execute(new SimpleStatement(
         "ALTER ROLE sam WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'connections'}"));
         session.execute(new SimpleStatement(
-        "GRANT MODIFY ON ecks.ectbl TO sam"));
+        "GRANT MODIFY ON ecks_itva.ectbl TO sam"));
         session.execute(new SimpleStatement(
-        "GRANT SELECT ON ecks.ectbl TO sam"));
+        "GRANT SELECT ON ecks_itva.ectbl TO sam"));
 
         session.execute(new SimpleStatement(
         "CREATE ROLE foo WITH PASSWORD = 'secret' AND LOGIN = true AND SUPERUSER = true"));
@@ -176,14 +176,20 @@ public class ITVerifyAudit
     @AfterClass
     public static void afterClass()
     {
-        session.execute(new SimpleStatement("DROP KEYSPACE IF EXISTS ecks"));
-        session.execute(new SimpleStatement("DROP ROLE IF EXISTS ecuser"));
-        session.execute(new SimpleStatement("DROP ROLE IF EXISTS foo"));
-        session.execute(new SimpleStatement("DROP ROLE IF EXISTS bar"));
-        session.execute(new SimpleStatement("DROP ROLE IF EXISTS mute"));
+        if(session != null)
+        {
+            session.execute(new SimpleStatement("DROP KEYSPACE IF EXISTS ecks_itva"));
+            session.execute(new SimpleStatement("DROP ROLE IF EXISTS ecuser"));
+            session.execute(new SimpleStatement("DROP ROLE IF EXISTS foo"));
+            session.execute(new SimpleStatement("DROP ROLE IF EXISTS bar"));
+            session.execute(new SimpleStatement("DROP ROLE IF EXISTS mute"));
 
-        session.close();
-        cluster.close();
+            session.close();
+        }
+        if(cluster != null)
+        {
+            cluster.close();
+        }
     }
 
     @Test
@@ -291,18 +297,18 @@ public class ITVerifyAudit
     public void testValidPreparedStatementsAreLogged()
     {
         PreparedStatement preparedInsertStatement = session
-                .prepare("INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?)");
+                .prepare("INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?)");
         PreparedStatement preparedSelectStatement = session
-                .prepare("SELECT * FROM ecks.ectbl WHERE partk = ?");
-        PreparedStatement preparedDeleteStatement = session.prepare("DELETE FROM ecks.ectbl WHERE partk = ?");
+                .prepare("SELECT * FROM ecks_itva.ectbl WHERE partk = ?");
+        PreparedStatement preparedDeleteStatement = session.prepare("DELETE FROM ecks_itva.ectbl WHERE partk = ?");
 
         List<String> expectedStatements = Arrays.asList(
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?)[1, '1', 'valid']",
-                "SELECT * FROM ecks.ectbl WHERE partk = ?[1]",
-                "DELETE FROM ecks.ectbl WHERE partk = ?[1]",
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?)[2, '2', 'valid']",
-                "SELECT * FROM ecks.ectbl WHERE partk = ?[2]",
-                "DELETE FROM ecks.ectbl WHERE partk = ?[2]");
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?)[1, '1', 'valid']",
+                "SELECT * FROM ecks_itva.ectbl WHERE partk = ?[1]",
+                "DELETE FROM ecks_itva.ectbl WHERE partk = ?[1]",
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?)[2, '2', 'valid']",
+                "SELECT * FROM ecks_itva.ectbl WHERE partk = ?[2]",
+                "DELETE FROM ecks_itva.ectbl WHERE partk = ?[2]");
 
         for (int i = 1; i <= 2; i++)
         {
@@ -326,22 +332,22 @@ public class ITVerifyAudit
     public void testValidBatchStatementsAreLogged()
     {
         PreparedStatement preparedInsertStatement1 = session
-                .prepare("INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?)");
+                .prepare("INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?)");
         PreparedStatement preparedInsertStatement2 = session
-                .prepare("INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, 'valid')");
+                .prepare("INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, 'valid')");
 
         List<String> expectedStatements = Arrays.asList(
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?)[1, '1', 'valid']",
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?)[2, '2', 'valid']",
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, 'valid')[3, '3']",
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')",
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, 'valid')[5, '5']");
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?)[1, '1', 'valid']",
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?)[2, '2', 'valid']",
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, 'valid')[3, '3']",
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')",
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, 'valid')[5, '5']");
 
         BatchStatement batch = new BatchStatement(BatchStatement.Type.UNLOGGED);
         batch.add(preparedInsertStatement1.bind(1, "1", "valid"));
         batch.add(preparedInsertStatement1.bind(2, "2", "valid"));
         batch.add(preparedInsertStatement2.bind(3, "3"));
-        batch.add(new SimpleStatement("INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')"));
+        batch.add(new SimpleStatement("INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')"));
         batch.add(preparedInsertStatement2.bind(5, "5"));
         session.execute(batch);
 
@@ -361,11 +367,11 @@ public class ITVerifyAudit
     public void testValidNonPreparedBatchStatementsAreLogged()
     {
         List<String> allStatements = Arrays.asList(
-        "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (1, '1', 'valid')",
-        "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (2, '2', 'valid')",
-        "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (3, '3', 'valid')",
-        "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')",
-        "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (5, '5', 'valid')");
+        "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (1, '1', 'valid')",
+        "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (2, '2', 'valid')",
+        "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (3, '3', 'valid')",
+        "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')",
+        "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (5, '5', 'valid')");
 
         StringBuilder batchStatementBuilder = new StringBuilder("BEGIN UNLOGGED BATCH ");
         for (String statement : allStatements)
@@ -393,8 +399,8 @@ public class ITVerifyAudit
     public void testValidPreparedBatchStatementsAreLogged()
     {
         String statement = "BEGIN UNLOGGED BATCH USING TIMESTAMP ? " +
-                           "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, ?); " +
-                           "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (?, ?, 'valid'); " +
+                           "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, ?); " +
+                           "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (?, ?, 'valid'); " +
                            "APPLY BATCH;";
 
         List<String> expectedStatements = Collections.singletonList(statement + "[1234, 1, '1', 'valid', 3, '3']");
@@ -417,12 +423,12 @@ public class ITVerifyAudit
     public void testValidPreparedStatementTypesAreLogged() throws Exception
     {
         PreparedStatement preparedStatement = session
-                .prepare("INSERT INTO ecks.ectypetbl "
+                .prepare("INSERT INTO ecks_itva.ectypetbl "
                         + "(partk, v0, v1, v2, v4, v5, v9, v13, v15)"
                         + " VALUES "
                         + "(?, ?, ?, ?, ?, ?, ?, ?, ?)");
 
-        String expectedStatement = "INSERT INTO ecks.ectypetbl "
+        String expectedStatement = "INSERT INTO ecks_itva.ectypetbl "
                 + "(partk, v0, v1, v2, v4, v5, v9, v13, v15)"
                 + " VALUES "
                 + "(?, ?, ?, ?, ?, ?, ?, ?, ?)[1, 'text', 'ascii', 123123123123123123, true, 1976-02-25, 8.8.8.8, 2004-05-29T14:29:00.000Z, 'varchar']";
@@ -449,7 +455,7 @@ public class ITVerifyAudit
     @Test
     public void testValidSimpleStatementTypesAreLogged()
     {
-        String statement = "INSERT INTO ecks.ectypetbl "
+        String statement = "INSERT INTO ecks_itva.ectypetbl "
                 + "(partk, v0, v1, v2, v4, v5, v9, v13, v15)"
                 + " VALUES "
                 + "(1, 'text', 'ascii', 123123123123123123, true, '1976-02-25', '8.8.8.8', '2004-05-29T14:29:00.000Z', 'varchar')";
@@ -501,9 +507,9 @@ public class ITVerifyAudit
         // Driver or Cassandra will add double-quotes to ks on one of the connections if statemens doesn't have it here.
         // TODO: Research if this is "bug" in Cassandra, driver or ecAudit?
         List<String> statements = Arrays.asList(
-                "INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (1, 'one', 'valid')",
-                "SELECT * FROM ecks.ectbl",
-                "USE \"ecks\"",
+                "INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (1, 'one', 'valid')",
+                "SELECT * FROM ecks_itva.ectbl",
+                "USE \"ecks_itva\"",
                 "INSERT INTO ectbl (partk, clustk, value) VALUES (1, 'one', 'valid')",
                 "SELECT * FROM ectbl");
 
@@ -527,7 +533,7 @@ public class ITVerifyAudit
                 .stream()
                 .map(ILoggingEvent::getFormattedMessage)
                 .collect(Collectors.toList()))
-                .containsOnlyElementsOf(expectedAttemptsAsUser(Arrays.asList("USE \"ecks\""), user));
+                .containsOnlyElementsOf(expectedAttemptsAsUser(Arrays.asList("USE \"ecks_itva\""), user));
     }
 
     /**
@@ -544,7 +550,7 @@ public class ITVerifyAudit
         try (Cluster privateCluster = cdt.createCluster(user, "secret");
              Session privateSession = privateCluster.connect())
         {
-            executeOneUseWithFollowingSelect(user, privateSession, "USE \"ecks\"");
+            executeOneUseWithFollowingSelect(user, privateSession, "USE \"ecks_itva\"");
             executeOneUseWithFollowingSelect(user, privateSession, "USE \"ecks2\"");
             executeOneUseWithFollowingSelect(user, privateSession, "USE \"ecks3\"");
         }
@@ -553,14 +559,14 @@ public class ITVerifyAudit
     private void executeOneUseWithFollowingSelect(String user, Session privateSession, String useStatement) {
         ArgumentCaptor<ILoggingEvent> loggingEventCaptor1 = ArgumentCaptor.forClass(ILoggingEvent.class);
         privateSession.execute(new SimpleStatement(useStatement));
-        privateSession.execute(new SimpleStatement("SELECT * FROM ecks.ectypetbl"));
+        privateSession.execute(new SimpleStatement("SELECT * FROM ecks_itva.ectypetbl"));
         verify(mockAuditAppender, atLeast(2)).doAppend(loggingEventCaptor1.capture());
         List<ILoggingEvent> loggingEvents1 = loggingEventCaptor1.getAllValues();
         assertThat(loggingEvents1
                 .stream()
                 .map(ILoggingEvent::getFormattedMessage)
                 .collect(Collectors.toList()))
-                .containsOnlyElementsOf(expectedAttemptsAsUser(Arrays.asList(useStatement, "SELECT * FROM ecks.ectypetbl"), user));
+                .containsOnlyElementsOf(expectedAttemptsAsUser(Arrays.asList(useStatement, "SELECT * FROM ecks_itva.ectypetbl"), user));
         reset(mockAuditAppender);
     }
 
@@ -569,11 +575,11 @@ public class ITVerifyAudit
     {
         List<String> statements = Arrays.asList(
                 "CREATE ROLE ecuser WITH PASSWORD = 'secret' AND LOGIN = true",
-                "CREATE KEYSPACE ecks WITH REPLICATION = {'class' : 'SimpleStrategy', 'replication_factor' : 1} AND DURABLE_WRITES = false",
-                "CREATE TABLE ecks.ectbl (partk int PRIMARY KEY, clustk text, value text)",
+                "CREATE KEYSPACE ecks_itva WITH REPLICATION = {'class' : 'SimpleStrategy', 'replication_factor' : 1} AND DURABLE_WRITES = false",
+                "CREATE TABLE ecks_itva.ectbl (partk int PRIMARY KEY, clustk text, value text)",
                 "INSERT INTO invalidks.invalidtbl (partk, clustk, value) VALUES (1, 'one', 'valid')",
                 "SELECT * FROM invalidks.invalidtbl",
-                "SELECT * FROM ecks.invalidtbl",
+                "SELECT * FROM ecks_itva.invalidtbl",
                 "DELETE FROM invalidks.invalidtbl WHERE partk = 2",
                 "DROP KEYSPACE invalidks",
                 "DROP ROLE invaliduser",
@@ -600,9 +606,9 @@ public class ITVerifyAudit
     {
         List<String> statements = Arrays.asList(
                 "SELECT * FROM nonexistingks.nonexistingtbl",
-                "SELECT * FROM ecks.nonexistingtbl",
+                "SELECT * FROM ecks_itva.nonexistingtbl",
                 "INSERT INTO nonexistingks.nonexistingtbl (partk, clustk, value) VALUES (1, 'one', 'valid')",
-                "INSERT INTO ecks.nonexistingtbl (partk, clustk, value) VALUES (1, 'one', 'valid')");
+                "INSERT INTO ecks_itva.nonexistingtbl (partk, clustk, value) VALUES (1, 'one', 'valid')");
 
         try (Cluster privateCluster = cdt.createCluster("sam", "secret");
              Session privateSession = privateCluster.connect())
@@ -627,7 +633,7 @@ public class ITVerifyAudit
             for (String statement : statements)
             {
                 BatchStatement batch = new BatchStatement(BatchStatement.Type.UNLOGGED);
-                batch.add(new SimpleStatement("INSERT INTO ecks.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')"));
+                batch.add(new SimpleStatement("INSERT INTO ecks_itva.ectbl (partk, clustk, value) VALUES (4, '4', 'valid')"));
                 batch.add(new SimpleStatement(statement));
                 assertThatExceptionOfType(InvalidQueryException.class).isThrownBy(() -> privateSession.execute(batch));
             }
